TCPSocket in parent process
